fixed provisioning
